102-fibonacci.c Final ISA

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,9 +7,9 @@
 int main(void)
 {
 int i;
-int num;
-int value;
-int tmp;
+unsigned long num;
+unsigned long value;
+unsigned long tmp;
 
 num = 0;
 value = 1;
